openapi_3 to swagger_2: Fix invalid body parameter
An OAS v3 spec with an operation to send a file in a request body, as
described in [Considerations for File Uploads][1]:

    requestBody:
      content:
        application/octet-stream:
          schema:
            type: string
            format: binary

api-spec-converter converts this to:

    consumes:
      - application/octet-stream
    parameters:
      - in: formData
        name: body
        type: file
    responses:
      '204':
        description: Success
    operationId: uploadFile

Which is invalid because ["if type is `file`, the consumes MUST be either
`multipart/form-data`, `application/x-www-form-urlencoded` or both"][2].

Instead, preserve the schema from the OAS3 spec, if present, or use
`type: "string"` with `format: "binary"` as recommended in
OAI/OpenAPI-Specification#326.  Also, don't copy the non-schema
properties, which are invalid for `in: "body"`.

[1]: https://github.com/OAI/OpenAPI-Specification/blob/master/versions/3.0.2.md#considerations-for-file-uploads
[2]: https://github.com/OAI/OpenAPI-Specification/blob/master/versions/2.0.md#fixed-fields-7

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>