Improve discriminator conversion for openapi_3 to swagger_2 (#206)
* openapi_3 to swagger_2: Discriminator name

In Swagger 2, the `discriminator` property has a string value with "the
schema property name that is used to differentiate between other schema
that inherit this schema".  In OpenAPI 3 the discriminator property is a
Discriminator Object with a `propertyName` property that corresponds to
the Swagger 2 definition.  Add the conversion.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>

* openapi_3 to swagger_2: discriminator mapping

OpenAPI 3 introduced a way to map property values to schema names called
a discriminator mapping.  Although Swagger 2 did not support this
mapping, some code generators support extensions which serve the same
purpose.  Convert `discriminator.mapping` to `x-discriminator-value`
(for Swagger Codegen) and `x-ms-discriminator-value` (for AutoRest) on
the discriminated schema objects.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>