Fix openapi_3 discriminator mapping to swagger_2
The OpenAPI 3 Specification defines the mapping property of the
Discriminator object as ["An object to hold mappings between payload
values and schema names or references."][1]  When the mapping contained a
schema name, the previous code (that I wrote, whoops) would crash with

    Error: Remote $ref URLs are not currently supported for openapi_3

Update the code to correctly handle schema names and to handle
resolution errors more gracefully.  Update tests to cover this.

Ref: #206

[1]: https://github.com/OAI/OpenAPI-Specification/blob/master/versions/3.0.2.md#discriminator-object

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>