openapi_3 to 2: formData as schema for name+in
When converting an OpenAPI 3 document to OpenAPI 2, if the schema for a
formData parameter is a reference, the name and in properties were added
to the $ref object, then lost when it was resolved.

This commit adds a test case to cover this (using an enum) and avoids
the issue by passing the schema as the schema property of a new
parameter object with the name and in properties.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>