openapi_3 to 2: Copy schema description to param
If the parameter doesn't have a description and the schema (possibly
from a property) does, copy the description to provide information about
the parameter.

This is intended to be both useful for users and to preserve the current
behavior of formData properties when they are passed as the schema
property instead of the parameter object in a subsequent commit.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>