openapi_3 to 2: Copy x- props from schema to param
The parameter in OpenAPI/Swagger 2 has the properties of schema,
excluding properties which define objects, and acts like schema.  When
translating, copy schema extension properties which are applicable to
parameters.  Since I'm not aware of any which are incompatible, copy
them all for now.

Add x-ms-enum to test.  Note that Autorest uses the same generated enum
type for multiple parameters (and schemas) with the same x-ms-enum.name
and values.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>