Resolve multiple identical refs in a single file
Resolves issue #248, and adds the testcase mentioned in the ticket as the test for this change.

This changes the resolve mechanism to clone the resolved entity before returning it, so that changes applied to it in the code that is using it, won't affect the component it references (and thus will not modify it and prevent it from being resolved in later occasions).

This change makes the x-component entities look like they originally were (unmodified), and this complies with the swagger 2 syntax, since their definition resides inside the extension property (see https://github.com/OAI/OpenAPI-Specification/blob/master/versions/2.0.md#specification-extensions).